Upgrade dependencies and fix demo.py

+++ llava/model/multimodal_encoder/clip_encoder.py
@@ -27,7 +27,12 @@ class CLIPVisionTower(nn.Module):
             return
 
         self.image_processor = CLIPImageProcessor.from_pretrained(self.vision_tower_name)
-        self.vision_tower = CLIPVisionModel.from_pretrained(self.vision_tower_name, device_map=device_map)
+        # Load vision tower with fp16 to save memory
+        self.vision_tower = CLIPVisionModel.from_pretrained(
+            self.vision_tower_name,
+            device_map=device_map,
+            torch_dtype=torch.float16
+        )
         self.vision_tower.requires_grad_(False)
 
         self.is_loaded = True
@@ -116,7 +121,12 @@ class CLIPVisionTowerS2(CLIPVisionTower):
             return
 
         self.image_processor = CLIPImageProcessor.from_pretrained(self.vision_tower_name)
-        self.vision_tower = CLIPVisionModel.from_pretrained(self.vision_tower_name, device_map=device_map)
+        # Load vision tower with fp16 to save memory
+        self.vision_tower = CLIPVisionModel.from_pretrained(
+            self.vision_tower_name,
+            device_map=device_map,
+            torch_dtype=torch.float16
+        )
         self.vision_tower.requires_grad_(False)
 
         self.image_processor.size['shortest_edge'] = self.s2_image_size
